// src/types/forms/formRefs.ts
export type FormRefs = {
  name: React.MutableRefObject<HTMLInputElement>;
  age: React.MutableRefObject<HTMLInputElement>;
  email: React.MutableRefObject<HTMLInputElement>;
  gender: React.MutableRefObject<HTMLSelectElement>;
  pass: React.MutableRefObject<HTMLInputElement>;
  confirmPass: React.MutableRefObject<HTMLInputElement>;
  ts: React.MutableRefObject<HTMLInputElement>;
  file: React.MutableRefObject<HTMLInputElement>;
  country: React.MutableRefObject<HTMLInputElement>;
};